working on agent system

+++ database/migrations/2017_07_16_015507_create_pc_user_amazon_items_table.php
@@ -6,34 +6,39 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
 class CreatePcUserAmazonItemsTable extends Migration {
+
+    public $tableName = 'pc_user_amazon_items';
+
     /**
      * Run the migrations.
      *
      * @return void
      */
     public function up() {
-        Schema::enableForeignKeyConstraints();
+        if (!Schema::hasTable($this->tableName)) {
+            Schema::disableForeignKeyConstraints();
+            Schema::create($this->tableName, function (Blueprint $table) {
+                $table->increments('id')->unsigned();
+                $table->unsignedInteger('user_id');
+                $table->unsignedInteger('pc_amazon_item_id');
 
-        Schema::create('pc_user_amazon_items', function (Blueprint $table) {
-            $table->increments('id')->unsigned();
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('pc_amazon_item_id');
+                $table->dateTime('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
+                $table->dateTime('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
 
-            $table->dateTime('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
-            $table->dateTime('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
+                $table->index(['user_id', 'pc_amazon_item_id']);
 
-            $table->index(['user_id', 'pc_amazon_item_id']);
+                $table->foreign('user_id')
+                    ->references('id')->on('users')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
 
-            $table->foreign('user_id')
-                ->references('id')->on('users')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-
-            $table->foreign('pc_amazon_item_id')
-                ->references('id')->on('pc_amazon_items')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-        });
+                $table->foreign('pc_amazon_item_id')
+                    ->references('id')->on('pc_amazon_items')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            });
+            Schema::enableForeignKeyConstraints();
+        }
     }
 
     /**
@@ -42,6 +47,6 @@ class CreatePcUserAmazonItemsTable extends Migration {
      * @return void
      */
     public function down() {
-        Schema::dropIfExists('pc_user_amazon_items');
+        Schema::dropIfExists($this->tableName);
     }
 }
